Move factory module out of services
This will help distinguish it from the services.
